@@ -7,25 +7,24 @@
   "homepage": "https://learn-reactjs.netlify.com/",
   "main": "src/index.js",
   "dependencies": {
-    "jest-dom": "^1.12.0",
-    "prop-types": "^15.6.2",
-    "react": "^16.4.2",
-    "react-dom": "^16.4.2",
-    "react-router": "^4.3.1",
-    "react-router-dom": "^4.3.1",
-    "react-testing-library": "^5.0.0",
-    "vanilla-tilt": "^1.4.1"
+    "jest-dom": "3.1.2",
+    "prop-types": "15.7.2",
+    "react": "16.8.2",
+    "react-dom": "16.8.2",
+    "react-router": "4.3.1",
+    "react-router-dom": "4.3.1",
+    "react-testing-library": "5.9.0",
+    "vanilla-tilt": "1.6.1"
   },
   "devDependencies": {
-    "all-contributors-cli": "^5.4.0",
-    "cross-spawn": "^6.0.5",
-    "husky": "^0.14.3",
-    "inquirer": "^6.1.0",
-    "is-ci": "^1.1.0",
-    "node": "^8.11.1",
-    "npm-run-all": "^4.1.3",
-    "react-scripts": "2.0.0-next.3e165448",
-    "replace-in-file": "^3.4.2"
+    "cross-spawn": "6.0.5",
+    "husky": "1.3.1",
+    "inquirer": "6.2.2",
+    "is-ci": "2.0.0",
+    "node": "10.15.1",
+    "npm-run-all": "4.1.5",
+    "react-scripts": "2.1.5",
+    "replace-in-file": "3.4.3"
   },
   "scripts": {
     "start": "react-scripts start",
